Refactor conda-forge recipe and implement aiida meta-package
This commit refactors the current conda-forge recipe for aiida-core
In this refactoring the previous aiida-core package was changed to
be a true sub-package of this recipe with all dependencies being
moved to the corresponding aiida-core sub-package located in the
outputs section. Additionally, a new metapackage "aiida" was
defined which acts as the top-level package name. This metapackage
is now used to bundle a selected set of sub-packages, defined by this
recipe, into an aiida-metapackage which may be used to install
aiida-core and required dependencies in one go